% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prorepparcodes.R
\name{addproreppar2tbl}
\alias{addproreppar2tbl}
\title{Add product reporter and partner to a tbl object}
\usage{
addproreppar2tbl(RMySQLcon, maintbl)
}
\arguments{
\item{RMySQLcon}{database connection object created by RMySQL \code{\link[DBI]{dbConnect}}}

\item{maintbl}{tbl containing trade data, with productcode, reportercode and partnercode}
}
\value{
a tbl object left joined to the product, reporter and partner tables.
}
\description{
Add product reporter and partner to a tbl object
}
\examples{
\dontrun{
con <- RMySQL::dbConnect(RMySQL::MySQL(), dbname = "test")
monthly <- tbl(con, "raw_comext_monthly_201707")
monthly \%>\%
    filter(productcode == 44) \%>\%
    addproreppar2tbl(con, .) \%>\%
    collect()
RMySQL::dbDisconnect(con)

}
}

