% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{cleancomextmonthly}
\alias{cleancomextmonthly}
\title{Clean Comext Monthly data}
\usage{
cleancomextmonthly(dtf)
}
\arguments{
\item{dtf}{data frame of monthly data for one product}
}
\description{
Re-uses several functions from the tradeflows package.
Each function mentionned specifically tradeflows::fun().
Hadley: http://r-pkgs.had.co.nz/namespace.html#imports
"
    It’s common for packages to be listed in Imports in DESCRIPTION, but not in NAMESPACE. 
    In fact, this is what I recommend: list the package in DESCRIPTION so that it’s installed,
    then always refer to it explicitly with pkg::fun().
"
}

