% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/product_code_changes.R
\docType{data}
\name{cn_code_changes}
\alias{cn_code_changes}
\title{CN code changes}
\format{A data frame with 15000 rows and 4 variables:
\describe{
  \item{origin_code}{CN code}
  \item{destination_code}{CN code}
  \item{orig_end_year}{year}
  \item{dest_start_year}{year}
}}
\source{
\url{https://ec.europa.eu/eurostat/ramon/relations/index.cfm?TargetUrl=LST_REL}
}
\usage{
cn_code_changes
}
\description{
This dataset is available as eutradeflows::cn_code_changes
Code changes are obtained from EUROSTAT Ramon
The Excel file is simply saved to a csv by hand.
}
\keyword{datasets}
