% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crontime.R
\name{crontime}
\alias{crontime}
\title{Test a cron job by writing the comext folder name to a text file}
\usage{
crontime(logfile, randomsleeptime)
}
\arguments{
\item{logfile}{character path to a log file}

\item{randomsleeptime}{integer maximum number of second to wait before checking the
comext website.}
}
\description{
This script waits a random number of seconds,
then scraps the most recent comext folder name and the name of the last file in that folder.
The ouput is written to a log file.
}
\details{
To run this script every day at 3:00 am, edit the crontab file

\code{
    sudo vim /etc/crontab
}

and enter the following command:

\code{
    0 2 * * *    debian Rscript -e "tradeharvester::crontime(logfile = '/mnt/sdb/public/log/crontime.txt', randomsleeptime = 3600)"
}

The log file can be located in a publicly accessible web folder.
}
\examples{
templog <- tempfile(fileext = ".txt")
crontime(logfile = templog, randomsleeptime = 0)
readLines(templog)
}
\seealso{
\code{\link{harvest}}, the main harvesting function run as a cron job.
}
