% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{dbconnecttradeflows}
\alias{dbconnecttradeflows}
\title{Create a database connection for Shiny}
\usage{
dbconnecttradeflows(dbdocker, dbname = "tradeflows")
}
\arguments{
\item{dbdocker}{logical specifying if the database connection is to be made with a docker container or not}
}
\value{
a database connection object, used to communicate with the database engine.
}
\description{
Depending on the dbdocker parameter, 
this function will return a database connection 
inside a docker container or a local database connection.
}
