% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prorepparcodes.R
\name{dbndistinct}
\alias{dbndistinct}
\title{Count the number of distinct rows in a database table for a given variable}
\usage{
dbndistinct(RMariaDBcon, tablename, variable)
}
\arguments{
\item{RMariaDBcon}{database connection object created by RMySQL \code{\link[DBI]{dbConnect}}}

\item{tablename}{character name of a database table}

\item{variable}{character name of a variable in that database table}
}
\value{
numeric value
}
\description{
Count the number of distinct rows in a database table for a given variable
}
\examples{
con <- RMariaDB::dbConnect(RMariaDB::MariaDB(), dbname = "test")
on.exit(RMariaDB::dbDisconnect(con))
# Transfer the iris data frame to the database
RMariaDB::dbWriteTable(con, "iris_in_db", iris, row.names = FALSE, overwrite = TRUE)
# Count the number of species
dbndistinct(con, "iris_in_db", Species)
}
