% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readcomext.R
\name{decompress7z}
\alias{decompress7z}
\title{Decompress a 7z file in its directory}
\usage{
decompress7z(
  file7z,
  stdout = FALSE,
  logfile = file.path("~/log", "harvesterrorlog.txt")
)
}
\arguments{
\item{file7z}{a 7z file to be decompressed}

\item{stdout}{redirects standard output of the 7zr command, argument passed to \link{system2}.
defaults to FALSE (discard output)}

\item{logfile}{path to a log file, located in the user directory by default}
}
\value{
Invisibly returns a character vector of the filepaths extracted to
 (same behaviour as \link{unzip})
return NULL if there is an error (the error is then logged to a file)
}
\description{
Overwrite by default, meaning
the 7zr switch is set to -y  "Assume Yes on all queries"
because there cannot be interaction with the user.
}
