% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadcomext.R
\name{downloadcomextfile}
\alias{downloadcomextfile}
\alias{downloadcomextmonthlyrecent}
\alias{downloadcomextmetadata}
\alias{downloadcomextmonthlyarchive}
\alias{downloadcomextyearlyarchive}
\title{Download bulk data from Comext into a designated folder}
\usage{
downloadcomextfile(
  comextfolder,
  comextfile,
  rawdatafolder,
  pause = 0,
  logfile = file.path("~/log", "harvesterrorlog.txt"),
  method = "libcurl"
)

downloadcomextmonthlyrecent(
  rawdatafolder,
  comextfolderpath = "/COMEXT_DATA/PRODUCTS",
  extension = ".7z",
  recentyears = 4,
  pause = 10
)

downloadcomextmetadata(
  rawdatafolder,
  comextfolderpath = "/COMEXT_METADATA/CLASSIFICATIONS_AND_RELATIONS/ENGLISH",
  extension = ".txt",
  recentyears = 4,
  pause = 10
)

downloadcomextmonthlyarchive(
  startyear,
  rawdatafolder,
  pause = 60,
  pattern = "S1\\\\]",
  extension = ".7z"
)

downloadcomextyearlyarchive(
  startyear,
  rawdatafolder,
  pattern = "S2\\\\]",
  extension = ".7z"
)
}
\arguments{
\item{comextfolder}{name of the folder starting with "comext/"}

\item{comextfile}{name of the file in the comext platform}

\item{rawdatafolder}{folder where the raw data will be storred}

\item{pause}{numeric pause time in seconds before downloading the file (usefull for multiple downloads)}

\item{logfile}{path to a log file, located in the user directory by default}

\item{method, }{see \code{\link{download.file}}}

\item{recentyears}{numeric number of years, will only download most recent files for
the given number of years. Use a large number (>20) to load all data the first time.}

\item{startyear}{numeric download files from that year onwards}

\item{pattern}{character pattern of the folder name, see \code{\link{scrapcomextfoldername}}}
}
\value{
\code{downloadcomextfile} Returns the download status, see \code{\link{download.file}}

\code{downloadcomextmonthlyrecent} returns a data frame
of file names and paths with their download status.

\code{downloadcomextmonthlyrecent} returns a data frame
of file names and paths with their download status.
}
\description{
Download files from the Comext bulk data repository
located at http://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?sort=1&dir=comext
Make sure that the output of paste0(comextfolder, comextfile) converted to html characters
returns an URL that has a matching file name on the bulk download repository.

If the file doesn't exist, Comext will still return a page,
with a table containing the following message:
"File **** does not exist or is not readable on the server".
In this case, the download status will be 0 (sucess) even though
the file failed to download.

If the destination folder is empty, download all files.
Otherwise, if there are files in the destination folder,
download only data from the past recentyears years.
}
\details{
Pause time was introduced because download returned an
error status when downloadingmany files in a row.
Pause time can be decided individually in the various functions that call
downloadcomextfile, usually a few seconds to a few minutes should be enough.
}
\examples{
\dontrun{
# This example will get outdated as time goes by
# Check the bulk download repository at
# http://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?sort=1&dir=comext
# for the current name of the comext most recent data folder
# Both should work,
# with trailing slash in the folder name
downloadcomextfile("comext/201706/data/", "nc201701.7z", "/tmp")
# without trailing slash in the folder name
downloadcomextfile("comext/201706/data", "nc201702.7z", "/tmp")
# Test error logging by downloading empty file in empty folder
downloadcomextfile("", "", "/tmp", logfile = "~/comextlog.txt")
}
\dontrun{
# Downloads all recent .7z data files into the /tmp folder
# and returns a dataframe with a status column
# describing the status of the download for each file
dtf <- downloadcomextmonthlyrecent("/tmp")
dtf$status
# Download all recent .txt description files
# describing products and reporting countries into the /tmp folder
dtf2 <- downloadcomextmonthlyrecent("/tmp", subfolder = "text/english", extension = ".txt")
}
\dontrun{
downloadcomextmetadata(rawdatafolder = "/tmp", pause = 0)
}
\dontrun{
# download monthly archive from 2000
downloadcomextmonthlyarchive(startyear = 2000, rawdatafolder = "/tmp")
}
\dontrun{
# download yearly archive from 2000
downloadcomextyearlyarchive(startyear = 2000, rawdatafolder = "/tmp")
}
}
