% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{dumptable}
\alias{dumptable}
\title{Copy the content of a database table to a compressed SQL dump file}
\usage{
dumptable(databasename, tablename, dumpfolder = "/mnt/sdb/data_raw/sqldump/")
}
\arguments{
\item{databasename}{character name of the database}

\item{tablename}{character name of the table}

\item{dumpfolder}{character path where the .sql.7z file will be saved,
defaults to a folder on the server, which might not exist on your system.}
}
\description{
This funcion uses the following shell utilities:
the mysqldump program to dump a database table.
the 7zr program to compress the resulting .sql text file into a .sql.7z file.
If mysqldump and 7zr are not available on your system, this function 
will not work.
}
\examples{
\dontrun{ # Dump a database table
dumptable("tradeflows", "raw_comext_monthly_201709")
}
}
\seealso{
\code{\link{loadtabledump}}
}
