% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eutradeflows.R
\docType{package}
\name{eutradeflows-package}
\alias{eutradeflows}
\alias{eutradeflows-package}
\title{eutradeflows: A validation algorithm for forest products trade data from Comext}
\description{
A Validation Algorithm for Forest Products Trade Data from Comext
}
\author{
\strong{Maintainer}: Paul Rougieux \email{paul.rougieux@gmail.com}

}
