% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readcomext.R
\name{filterandrenamedat}
\alias{filterandrenamedat}
\title{Create a shorter .dat file containing only products starting with the given
code and rename colmuns to an internal naming scheme.}
\usage{
filterandrenamedat(datfile, productcodestart)
}
\arguments{
\item{datfile}{character path to a .dat file}

\item{productcodestart}{numeric or character "08" product code starting with}
}
\value{
character name of the result file
}
\description{
Prepare a .dat file for its transfer into the database.
See docs/read_comext_files.Rmd for developments that let to this awk command:
\code{
awk -F, 'NR==1 {print "different column"} $3 ~ /^44/' nc201501.dat|less
}

Prepare a dat file for input into the database
}
\examples{
\dontrun{
# Download a file from the [comext bulk download repository](http://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?sort=1&dir=comext), then
mkdir /tmp/comext
then extract the .7z file with the bash command
# 7zr e nc201501.7z
# then prepare the .dat file with the R command
filterandrenamedat("/tmp/comext/nc201501.dat",44)
}
}
