% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadcomext.R
\name{generatemonthlyfilenames}
\alias{generatemonthlyfilenames}
\alias{generateyearlyfilenames}
\title{Generate all monthly file names from startyear until endyear}
\usage{
generatemonthlyfilenames(
  startyear,
  endyear = as.integer(format(Sys.Date(), "\%Y")),
  extension = ".7z"
)

generateyearlyfilenames(
  startyear,
  endyear = as.integer(format(Sys.Date(), "\%Y")),
  extension = ".7z"
)
}
\arguments{
\item{startyear}{numeric or character, a start year}

\item{endyear}{numeric or character, an end year, defaults to the current year}

\item{extension}{character vector of length one only, a file extension}
}
\value{
a vector of file names
}
\description{
Generate all monthly file names from startyear until endyear

Generate yearly file names
}
