% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest.R
\name{harvestrecent}
\alias{harvestrecent}
\alias{harvestmonthlyarchive}
\alias{harvestyearlyarchive}
\title{Deprecated harvest recent data
 @export}
\usage{
harvestrecent()

harvestmonthlyarchive(
  ...,
  rawdatafolder,
  startyear,
  tabletemplate = "raw_comext_monthly_template",
  template = getOption("tradeharvester")$template
)

harvestyearlyarchive(
  ...,
  rawdatafolder,
  startyear,
  tabletemplate = "raw_comext_yearly_template",
  template = getOption("tradeharvester")$template
)
}
\arguments{
\item{...}{parameters passed to \code{\link{transfer7zfolder2db}}}
}
\description{
\code{harvestmonthlyarchive}
Downloads archive data from comext and transfer all sub products
of the given product codes to the database.
The structure of the table template `raw_comext_monthly`
should already be present in the database.
The database structure is regularly recreated in \code{harvestrecent},
so there is no need to recreate it \code{harvestmonhtlyarchive}.
This function can be run on a yearly basis when new archives become available.

\code{harvestyearlyarchive}
downloads yearly archive and transfer it to the database.
The structure of the database table `raw_comext_yearly`
should already present in the database.
}
