% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{loadtabledump}
\alias{loadtabledump}
\title{Load a dump file into the database
This function uses shell utilities to load a sql dump into the database.}
\usage{
loadtabledump(databasename, dumpfile)
}
\arguments{
\item{databasename}{character name of the database}

\item{dumpfile}{character path to a dumpfile}
}
\description{
Load a dump file into the database
This function uses shell utilities to load a sql dump into the database.
}

