% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longest_contiguous.R
\name{longest_contiguous}
\alias{longest_contiguous}
\title{Longest contiguous time series for the given variable in the given data
frame}
\usage{
longest_contiguous(df, variable)
}
\arguments{
\item{df}{data frame}

\item{variable}{character name of the variable of interest}
}
\value{
list of start and end dates for the longest time series
}
\description{
Longest contiguous time series for the given variable in the given data
frame
}
