% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D3.R
\name{plotsankey}
\alias{plotsankey}
\title{Plot a sankey diagram, using the D3 library networkD3}
\usage{
plotsankey(linksandnodes, value = "tradevalue", units = "K€")
}
\arguments{
\item{linksandnodes}{a list containing 2 dataframes: links and nodes}

\item{value}{the value to be displayed in the sankey diagram, a column 
in linksandnodes$links}
}
\description{
Plot a sankey diagram, using the D3 library networkD3
}
\examples{
\dontrun{
con <- RMySQL::dbConnect(RMySQL::MySQL(), dbname = "tradeflows")
wood <- tbl(con, "vld_comext_monthly") \%>\% 
   filter(productcode ==  44072199 & 
              flowcode == 1 & 
              period == "201708") \%>\% 
   # head(200) \%>\% 
   collect()
reporter <- tbl(con, "vld_comext_reporter") \%>\% collect()
partner <- tbl(con, "vld_comext_partner") \%>\% collect()
wood \%>\% 
    preparesankeynodes \%>\% 
    plotsankey()
}
}
