% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D3.R
\name{preparesankeynodes}
\alias{preparesankeynodes}
\title{Prepare nodes and links to generate a Sankey diagram}
\usage{
preparesankeynodes(dtf, reporter, partner, debugname = FALSE)
}
\arguments{
\item{dtf}{a data frame of trade flows data}

\item{reporter}{data frame of reporter codes and names}

\item{partner}{data frame of partner codes and names}
}
\value{
a list of data frames with links and nodes
}
\description{
Prepare nodes and links to generate a Sankey diagram
}
