% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/product_code_changes.R
\name{previous_codes}
\alias{previous_codes}
\title{Function to find previous codes}
\usage{
previous_codes(
  dest_codes_to_check,
  cn_code_changes = eutradeflows::cn_code_changes
)
}
\arguments{
\item{dest_codes_to_check}{usually only one code}
}
\description{
Function to find previous codes
}
\examples{
previous_codes(cn_code_changes, "44072969")
}
