% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readcomext.R
\name{readdat}
\alias{readdat}
\title{Read Comext .dat file for a given product code}
\usage{
readdat(datfile, productcodestart)
}
\arguments{
\item{datfile}{character path to the .dat file}

\item{productcodestart}{numeric begining of a comext product code}
}
\description{
This function uses the awk program to filter large .dat file
before reading them as csv.
This is a convenience function for raw data exploration,
the transfer to the database itself is performed by another function
called \code{\link{filterandrenamedat}}.
}
\seealso{
write2db for direct transfer of the dat file to the database
}
