% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_tables.R
\name{replace_repeated_by_empty_string}
\alias{replace_repeated_by_empty_string}
\title{Replace repeated values by the empty string}
\usage{
replace_repeated_by_empty_string(x, empty_value = "")
}
\arguments{
\item{x}{a vector}

\item{empty_value}{the default values used for the lag, should be different than the first value in x}
}
\value{
vector
}
\description{
This function replaces duplicated values in a column by the empty string so that the table output is easier to read.
The removal is done by comparing the vector with a shifted (lagged) version of itself.
By default this function will convert numeric ouptut to character variables, 
unless you change the default \code{empty_value=''} argument.
}
\examples{
replace_repeated_by_empty_string(c("bla","bla","bli", "bli"))
replace_repeated_by_empty_string(c(1,2,3))
replace_repeated_by_empty_string(c(0,0,0,1,2,2,3,3,3))
}
