% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_main_partners.R
\name{select_main_partners}
\alias{select_main_partners}
\title{A function to select main partner(s) for each reporter}
\usage{
select_main_partners(df, slice_n = 1)
}
\arguments{
\item{df}{a data frame of trade flows}

\item{slice_n}{number of partner(s) to slice, defaults to one}
}
\value{
a data frame with the main partner for each reporter
}
\description{
A function to select main partner(s) for each reporter
}
\examples{
\dontrun{
library(dplyr)
# Load sawnwood oak data from the database
con <- RMariaDB::dbConnect(RMariaDB::MariaDB(), dbname = "tradeflows")
swd_oak <- tbl(con, "raw_comext_monthly") \%>\%
    # Filter import flows and oak sawnwood
    filter(flowcode == 1 & productcode == "44079190") \%>\% 
    collect()
# Load reporter country codes and names
reporter_names <- tbl(con, 'vld_comext_reporter') \%>\% collect()
# Load partner codes and names
partner_names <- tbl(con, 'vld_comext_partner') \%>\%collect()
# Add reporter and partner information
swd_oak <- swd_oak \%>\%
    left_join(partner_names, by='partnercode') \%>\%
    left_join(reporter_names, by='reportercode')
RMariaDB::dbDisconnect(con)
# Now use this function:
swd_oak \%>\%
    select_main_partners()
}
}
