% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_main_partners.R
\name{select_main_partners_trop}
\alias{select_main_partners_trop}
\title{Select main tropical partners for the product of interest and related products}
\usage{
select_main_partners_trop(
  wood_imports,
  partner_codes_of_interest,
  product_codes_of_interest,
  force_full_series = TRUE
)
}
\arguments{
\item{wood_imports}{data frame of wood imports from comext}

\item{force_full_series}{boolean to force selection of only country pairs which have all data points for all period}

\item{vpa_partners}{data frame containing the country code of VPA partners}
}
\value{
data fram of main tropical sawnwood partners
}
\description{
Select main tropical partners for the product of interest and related products
}
\examples{
# Load the main partners of tropical sawnwood
swd_trop_main <- select_main_partners(wood_imports, vpa_partners, "44072969")
# Load the main partners of plywood
ply_trop_main <- select_main_partners(wood_imports, vpa_partners, "44123190")
}
