% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadcomext.R
\name{setcomextconfig}
\alias{setcomextconfig}
\title{Sets comext url connection parameters as a global R option comext}
\usage{
setcomextconfig(reload = FALSE, silent = FALSE)
}
\arguments{
\item{reload}{logical, force reloading the configuration file}

\item{silent}{logical, do not print "already loaded" message when TRUE}
}
\description{
Comext url parameters
Are storred in the installed version of the package
Run this command to find the location of the configuration file on your machine:
system.file("config/comextconfig.R", package="tradeharvester")
Use reload=TRUE to force reloading the file after modification.
}
\examples{
setcomextconfig(reload=TRUE)
comext <- getOption("comext")
comext["urlbulkfile"]
}
