% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/product_code_changes.R
\name{subsequent_codes}
\alias{subsequent_codes}
\title{Find subsequent code for a given product code}
\usage{
subsequent_codes(
  orig_codes_to_check,
  cn_code_changes = eutradeflows::cn_code_changes
)
}
\arguments{
\item{orig_codes_to_check}{character, usually one code}

\item{cn_code_changes}{a table of cn_code_changes (already contained in the package)}
}
\value{
data frame of all subsequent codes for the given original code
}
\description{
Find subsequent code for a given product code
}
\examples{
subsequent_codes(cn_code_changes, "44072969")
}
