% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_write.R
\name{transfer7zfile2db}
\alias{transfer7zfile2db}
\alias{transfer7zfolder2db}
\alias{transfertxtcodesfolder2db}
\title{Transfer data from .7z files or text files to the database}
\usage{
transfer7zfile2db(RMariaDBcon, file7z, productcodestart, tablename)

transfer7zfolder2db(
  RMariaDBcon,
  rawdatacomextfolder,
  productcodestart,
  tablename,
  tabletemplate
)

transfertxtcodesfolder2db(RMariaDBcon, rawdatacomextfolder)
}
\arguments{
\item{RMariaDBcon}{database connection object created by RMariaDB::\link{dbConnect}}

\item{file7z}{character path to a .7z file from Comext}

\item{productcodestart}{numeric vector of product codes. These products and
all their sub-products will be transfered to the database.
\itemize{
 \item{}{\code{transfer7zfile2db} transfers only one product code}
 \item{}{\code{transfer7zfolder2db} can transfer a vector of product codes}
 }}

\item{tablename}{character name of a database table}

\item{tabletemplate}{character name of a table template to be copied and
renamed to tablename}

\item{rawdatacomextfoler}{character path of the folder where
comext files have been saved localy (usually ends with "[year][month]/data")}
}
\description{
\code{transfer7zfile2db} transfers data from a .7z file to the database.

\code{transfer7zfolder2db} lists all .7z files available in the
local copy of the raw Comext folder and transfers them to a database table.
Attention: \code{transfer7zfolder2db} erases all previous content in the database.

\code{transfertxtcodesfolder2db} transfers product and country codes to the database
}
\details{
\code{transfer7zfile2db} is a chain of several functions that:
\itemize{
 \item{}{Decompress comext file .7z file into .dat file in a temporary directory}
 \item{}{filter .dat file, keep only products starting by a certain code, such as "44"}
 \item{}{rename columns from comext to efi column names}
 \item{}{write file content to a database table}
 \item{}{delete decompressed version of the comext file}
 \item{}{(delete temporary directory)}
}
}
\examples{
\dontrun{ # Transfer data from a 7z file to the database
# Create a database connection
con <- RMariaDB::dbConnect(RMariaDB::MariaDB(), dbname = "test")
# Transfer data for product codes 44 and 94 from a 7z file to a database table
transfer7zfile2db(con, "data_raw/comext/201706/data/nc201501.7z", c(44,94), "raw_comext_monthly_201706")

# Transfer data from many .7z files to the database
# decompose
harvestrecent(con, "/tmp",c(44,94))
# in several steps.
# Download the latest comext data
comextfiles <- downloadcomextmonthlyrecent("/tmp")
comextfolder <- unique(comextfiles$folder)
# Create a table name where this data will be saved
tablename <- paste0("raw_comext_monthly_",
                   scrapcomextfoldername(format(Sys.Date(),"\\\\[\%Y")))
# Transfer recent monthly data to the database
transfer7zfolder2db(con,
                    rawdatacomextfolder = file.path("/tmp",comextfolder),
                    productcodestart = c(44, 94),
                    tablename = tablename,
                    tabletemplate = "raw_comext_monthly_template")

# Transfer monthly archive to the database
transfer7zfolder2db(con,
                    rawdatacomextfolder = file.path("data_raw/comext/2016S1/data"),
                    productcodestart = c(44, 94),
                    tablename = "raw_comext_monthly_2016S1",
                    tabletemplate = "raw_comext_monthly_template")

# Transfer yearly archive to the database
# This time the template is raw_comext yearly
transfer7zfolder2db(con,
                    rawdatacomextfolder = file.path("data_raw/comext/2016S2/data"),
                    productcodestart = c(44, 94),
                    tablename = "raw_comext_yearly_2016S2",
                    tabletemplate = "raw_comext_yearly_template")

# Transfer product codes and country codes to the database
transfertxtcodesfolder2db(con,
                          rawdatacomextfolder = "data_raw/comext/201707/text/english/")

# Disconnect from the database
RMariaDB::dbDisconnect(con)
}
}
\seealso{
\code{\link{writedat2db}}

\code{\link{writetxtcodes2db}}
}
