% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/product_code_changes.R
\name{update_code_changes}
\alias{update_code_changes}
\title{Prepare the code changes table
Code changes are obtained from EUROSTAT Ramon
https://ec.europa.eu/eurostat/ramon/relations/index.cfm?TargetUrl=LST_REL
The Excel file is simply saved to a csv by hand.}
\usage{
update_code_changes(
  csv_file = "~/downloads/CN_2019_update_of_codes.csv",
  rdata_file = "data/cn_code_changes.rda"
)
}
\arguments{
\item{csv_file}{character path to the csv file from EUROSTAT Ramon where code changes are given}

\item{packagefodler}{character path to the rds file where to save the updated data. Defaults inside the package.}
}
\value{
a data frame of this form
 origin_code  destination_code  orig_end_year  dest_start_year
  <chr>       <chr>                    <dbl>           <dbl>
1 02012011    02012021                  1988            1989
2 02012011    02012029                  1988            1989
}
\description{
Prepare the code changes table
Code changes are obtained from EUROSTAT Ramon
https://ec.europa.eu/eurostat/ramon/relations/index.cfm?TargetUrl=LST_REL
The Excel file is simply saved to a csv by hand.
}
\examples{
\dontrun{
# Update the codes while developing the package
update_code_changes()
}
}
