% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{loadflows}
\alias{loadflows}
\title{Load trade flows for the given product}
\usage{
loadflows(
  RMariaDBcon,
  productanalysed,
  periodstart = (as.numeric(format(Sys.time(), "\%Y")) - 3) * 100,
  tableread = "vld_comext_monthly"
)
}
\description{
Load trade flows for the given product
}
