% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prorepparcodes.R
\name{addproreppar2tbl}
\alias{addproreppar2tbl}
\alias{addunit2tbl}
\title{Add product reporter and partner to a tbl object}
\usage{
addproreppar2tbl(RMariaDBcon, maintbl)

addunit2tbl(RMariaDBcon, maintbl, tableunit = "vld_comext_unit")
}
\arguments{
\item{RMariaDBcon}{database connection object created by RMySQL \code{\link[DBI]{dbConnect}}}

\item{maintbl}{tbl containing trade data, with productcode, reportercode and partnercode}
}
\value{
a tbl object left joined to the product, reporter and partner tables.
}
\description{
Add product reporter and partner to a tbl object
}
\details{
addunit2tbl joints all unit codes, 
then removes those which are out dated
}
\examples{
\dontrun{
con <- RMariaDB::dbConnect(RMariaDB::MariaDB(), dbname = "test")
monthly <- tbl(con, "raw_comext_monthly_201707")
monthly \%>\%
    filter(productcode == 44) \%>\%
    addproreppar2tbl(con, .) \%>\%
    collect()
RMariaDB::dbDisconnect(con)
}
}
