% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_main_partners.R
\name{select_main_partners}
\alias{select_main_partners}
\alias{select_wood_imports}
\alias{select_large_country_pairs}
\alias{display_table_percentage}
\alias{display_table_number_of_flows}
\title{Select main tropical partners for the product of interest and related products}
\usage{
select_main_partners(wood_imports, vpa_partners, product_code_of_interest)

select_wood_imports(wood_imports, vpa_partners, product_code_of_interest)

select_large_country_pairs(df, trade_pc_threshold = 0.01)

display_table_percentage(
  df_large_country_pairs,
 
    caption = "Percentage of trade value in the main country pairs compared to trade in all countries"
)

display_table_number_of_flows(
  df_large_country_pairs,
  caption = "Number of trade flows in the main country pairs"
)
}
\arguments{
\item{wood_imports}{data frame of wood imports from comext}

\item{vpa_partners}{data frame containing the country code of VPA partners}

\item{df}{a data frame of trade flows}

\item{df_large_country_pairs}{a data frame of main partners with columns `trade_pct` and `n`.
output of the select_large_country_pairs function.}

\item{caption}{character table caption}
}
\value{
data fram of main tropical sawnwood partners
}
\description{
Select the largest country pairs
Select all country pairs which represent at least a given
percentage (default to 1%) of the trade value over the whole period.
}
\examples{
# Load the main partners of tropical sawnwood
swd_trop_main <- select_main_partners(wood_imports, vpa_partners, "44072969")
# Load the main partners of plywood
ply_trop_main <- select_main_partners(wood_imports, vpa_partners, "44123190")
}
