% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{write2csv_spread}
\alias{write2csv_spread}
\alias{write2csv_long}
\title{Spread quantity data along the period and write to a csv file
Write a data frame containing trade flows to a csv file}
\usage{
write2csv_spread(dtf, csvfile)

write2csv_long(dtf, csvfile)
}
\arguments{
\item{dtf}{a data frame containing validated trade flows}

\item{csvfile}{character path to a csv file}
}
\value{
The input data frame invisibly
}
\description{
Spread quantity data along the period and write to a csv file
Write a data frame containing trade flows to a csv file
}
\examples{
\dontrun{ # Clean product and country codes
con <- RMySQL::dbConnect(RMySQL::MySQL(), dbname = "test")
csvfile <- tempfile(fileext = ".csv")
loadflows(con, productanalysed = "44071190", periodstart = 201700, 
          tableread = "vld_comext_monthly") \%>\% 
    write2csv_spread(csvfile)
readLines(csvfile,5) # Check the first lines of the csv file
RMySQL::dbDisconnect(con)
}
}
